Add new link component and seperate frames table into unique page
