+++ frontend/src/components/tables/genericTable.tsx
@@ -24,7 +24,7 @@ export const GenericTable = <T extends object>({
     <div className="flex flex-col">
       <div className="overflow-x-auto sm:-mx-6 lg:-mx-8">
         <div className="inline-block min-w-full py-4 sm:px-6 lg:px-8">
-          <div className="overflow-hidden p-2">
+          <div className="overflow-hidden py-2">
             <table className="min-w-full text-center">
               <thead className="border-b bg-gray-50">
                 {table.getHeaderGroups().map((headerGroup) => (
